backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.3
Pillow>=10.0.0
numpy>=1.24.0
